Upgrade to React Native 0.31 template
