+++ android/app/src/main/java/com/barteguiden/MainActivity.java
@@ -1,14 +1,6 @@
 package com.barteguiden;
 
 import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-import com.burnweb.rnsendintent.RNSendIntentPackage;
-import com.oblador.vectoricons.VectorIconsPackage;
-import cl.json.RNSharePackage;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class MainActivity extends ReactActivity {
 
@@ -21,26 +13,4 @@ public class MainActivity extends ReactActivity {
         return "Barteguiden";
     }
 
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            new RNSendIntentPackage(),
-            new VectorIconsPackage(),
-            new RNSharePackage()
-        );
-    }
 }
